[MODEL] BERT conversion scripts, SciBERT, BioBERT, ClinicalBERT (#735)
* Fix typos

* Refactor argparse

* Validate bert_config.json

* Ignore saved optimizer parameters

Typically Optimizer parameters are not exported for Bert models
(eg. https://github.com/google-research/bert). As this does not always hold, we
have to handle the case that the parameters are still present (eg.
https://github.com/allenai/scibert).

* Support non-standard datasets/vocabularies in compare_tf_gluon_model

* Make use of flexible vocab in scripts/bert/convert_tf_model.py

* Address comments

* Fix lint

* Move input.txt

* Fix vocabulary mismatch in compare_tf_gluon_model.py

* Add SciBert

* Add BioBert

* Add PyTorch Bert support

Confirmed that the scibert_scivocab_uncased loaded from PyTorch produces the
same output as the tensorflow version (based on the compare_tf_gluon_model.py).

* Fix lint

* Fix clinicalbert test case

* Address comments

* Rename biobert and clinicalbert models

* Update bert.py