[REFACTOR] Refactor TokenEmbedding to reduce number of places that initialize internals (#750)
Prior to this commit, the TokenEmbedding constructor could only construct an
empty TokenEmbedding. However, an empty TokenEmbedding is of little use, thus
there exist a variety of places that modify and overwrite TokenEmbedding
internals after construction to "fill" the idx_to_token, idx_to_vec. Examples
are _load_embedding_text or _load_embedding_serialized within the TokenEmbedding
class, but also set_embedding in the Vocab class.

This commits
1) makes these methods static and changes them to return the
   idx_to_token and idx_to_vec,
2) extends the TokenEmbedding constructor to allow constructing a "non-empty"
   TokenEmbedding given newly added idx_to_token and idx_to_vec arguments

This change is backwards compatible in that it does not change any public API
besides introducing idx_to_token and idx_to_vec arguments to TokenEmbedding. For
the future, the "empty" TokenEmbedding initialization may be removed as it
provides little benefit. For now it is kept for backwards compatibility.