[TUTORIAL][DOC] Tutorial Updates (#802)
* Updated word_embedding.md

Made grammar and flow changes; still have second half to-do

* Updated grammar and clarified unclear sections

Updated the grammar of several portions and clarified and broke up some code segments or explained longer code segments to make it more readable and beginner-friendly

* Updated word_embedding_training

Updated the order and added a section to remind the reader to download the archive before proceeding with the tutorial. English grammar mistakes, spelling mistakes, and other trivial errors were amended.

* Corrected grammar

Corrected grammar

* Updated erroneous update

* Update word embedding

changed

* Updated with changes from comments

Updated with the changes from the comments by Julian

* Updated language_model

Updated the tutorial for LSTM language models and added more explanations as well as code breaks with notes and comments in the code to allow for easier user understandability

* Updated Google Neural Machine Translation

Updated GNMT to be better documented with code breaks, comments, and links to outside sources for complex topics.

* Updated grammar and code blocks

Updated some grammar and code blocking, along with explanations for longer pieces of code

* Updated the self-attentive sentence embedding tutorial

Updated the tutorial for more clear and concise information, as well as making it more consistent with the previous tutorials and the mathematics formatted correctly.

* Updated bert.md

Updated bert fine-tuning script and the tutorial to be more cohesive and gramatically correct, and broke up large code segments into smaller chunks with better comments.

* Updated sentiment_analysis.md

Refactored code, made the tutorial more in-depth, and simplified some concepts as well as adding reference links for interested users.

* Updated sequence_sampling.md

Updated the file with better language, better formatting and better grammar. Added in some breaks in the code to allow for the reader to fully understand the tutorial's materials

* Update word_embedding.md

* fix code issues

* fix title level

* fix title level

* Updated with changes from aston

Updated with changes from aston

* Updated self_attentive_sentence_embedding.md

Changed the latex format that auto-centred to a better to read format for the single entries.

* Updated citations/references in unupdated notebooks

Updated citations/references title and standardized title depth

* Update sentiment_analysis.md

* Updated to hopefully not conflict with the tests

Updated titles/headers to hopefully not cause errors in the CI.

* Updated transformer

mistake in changing pretrain to pre-train