Change the variable model to bert_classifier
The "model" is a imported module from code section 1 and then overloaded on code section 4
```{.python}
model = model.classification.BERTClassifier(bert_base, num_classes=2, dropout=0.1)
```
This actually make this section not rerunable  and if anything need to call the module "model", it will call the BERTClassifier  object rather than the model imported in section 1. Propose to change the model to bert_classifier to avoid naming conflicts.