[Tests] Improve test robustness (#960)
* Make test_question_answering self-contained

nltk.download('punkt') inside the test to be independent of the test environment

* Change SpacyTokenizer default lang from 'en' to 'en_core_web_sm'

'en' is by default a alias of 'en_core_web_sm'. The aliasing logic of spacy is
not very robust, so in some situations spacy fails to detect the presence of
'en_core_web_sm' if only 'en' is specified.

* Consistently use tests/data as MXNET_HOME during test execution

In some test environments, $HOME may not be set / writable. Thus rely on
MXNET_HOME during test execution and configure it as the tests/data subdirectory
that is already used by convention (in most places).